+++ src/app/zerodha/_components/OrderForm/orderForm.tsx
@@ -6,8 +6,8 @@ import { z } from "zod";
 import { api } from "~/trpc/react";
 import { useToast } from "../../_contexts/Toast/toast-context";
 import BackndWSContext from "../../_contexts/backendWS/backendWS";
-import { updateFormData } from "../../redux/FormData/FormData";
-import { AppDispatch, RootState } from "../../redux/store";
+import { updateFormData } from "../../_redux/FormData/FormData";
+import { AppDispatch, RootState } from "../../_redux/store";
 import { TFormSchema } from "./FrmSchema";
 import InputDiv from "./InputDiv";
 import { CheckBox, OrderTypeDiv } from "./OrderTypeDiv";
