+++ src/app/zerodha/_redux/rightSideData/rightSideData.ts
