+++ src/app/zerodha/_redux/watchList/watchList.ts
